requirements.txt
aiogram==3.15.0
openai==1.54.3
httpx==0.27.0
python-dotenv>=1.0.0
Pillow>=10.0.0
asyncpg>=0.29.0
aiohttp>=3.8.0
APScheduler>=3.10.4
pytz>=2023.3
firebase-admin>=6.4.0
